Path finding is working on a simple exemple.

+++ src/lib/pathfindingmap.h
@@ -5,6 +5,7 @@
 // std
 #include <cmath>
 
+#include <QDebug>
 // core
 #include "generalmap.h" // not needed if we separate TilePos
 #include "grid.h"
@@ -22,12 +23,14 @@ public:
 
   double cost(const TilePos& from, const TilePos& to) const
   {
-    return std::pow(from.x - to.x, 2) + std::pow(from.y - to.y, 2);
+    return std::pow(double(from.x) - double(to.x), 2) +
+        std::pow(double(from.y) - double(to.y), 2);
   }
 
   double heuristic(const TilePos& pos, const TilePos& obj) const
   {
-    return std::pow(pos.x - obj.x, 2) + std::pow(pos.y - obj.y, 2);
+    return std::pow(double(pos.x) - double(obj.x), 2) +
+        std::pow(double(pos.y) - double(obj.y), 2);
   }
 
   const std::vector<TilePos>& neighbor(const TilePos& pos) const
